Shorten_link/requirements.txt
pyperclip==1.8.2
pyshorteners==1.0.1
Pillow==9.2.0
